requirements.txt
psycopg2==2.9.10
confluent-kafka==2.6.0
simplejson==3.19.3
numpy==2.0.2
pandas==2.2.3
requests=2.32.3
python-dotenv==1.0.1
pendulum==3.0.00
pyspark==3.5.3
streamlit==1.39.0
matplotlib==3.9.2
kafka-python==2.0.2
streamlit-autorefresh==1.0.1